updates to build and setup scripts

+++ setup_tools/buildlegolas.sh
@@ -1,9 +1,17 @@
-rm -rf build
-rm -f legolas
+if [ ! -f CMakeLists.txt ]; then
+    echo "CMakeLists.txt not found! Can not start build process."
+    exit
+fi
 if [ "$1" == "clean" ]
 then
-  echo "build directory removed."
+  rm -f legolas
   echo "legolas executable removed."
+  if [[ -z "${LEGOLASDIR}" ]]; then
+    echo "Environment variable \$LEGOLASDIR is not defined, can't clean."
+  else
+    rm -rf "$LEGOLASDIR/build"
+    echo "build directory removed."
+  fi
   exit
 fi
 mkdir build
@@ -11,3 +19,4 @@ cd build || exit
 cmake ..
 make -j 2
 cd ..
+rm -rf build
